Separate BranchInfo class into its own file
The global groovy script variables are instanciated on demand by
Jenkins, meaning that referring to BranchInfo in a file that's loaded
before common.groovy is referenced would result in a compilation error.

Hoisting the BranchInfo class into a separate file and explicitly
importing it resolves this issue.
