+++ vars/common.groovy
@@ -36,6 +36,8 @@ import hudson.AbortException
 import org.jenkinsci.plugins.github_branch_source.Connector
 import org.kohsuke.github.GHPermissionType
 
+import org.mbed.tls.jenkins.BranchInfo
+
 /* Indicates if CI is running on Open CI (hosted on https://ci.trustedfirmware.org/) */
 @Field is_open_ci_env = env.JENKINS_URL ==~ /\S+(trustedfirmware)\S+/
 
@@ -88,39 +90,6 @@ import org.kohsuke.github.GHPermissionType
 /* Maps platform names to the tag of the docker image used to test that platform.
  * Populated by init_docker_images() / gen_jobs.gen_dockerfile_builder_job(platform). */
 @Field static docker_tags = [:]
-
-
-class BranchInfo {
-    /* Map from component name to chosen platform to run it, or to null
-     * if no platform has been chosen yet. */
-    public Map<String, String> all_all_sh_components
-
-    /* Whether scripts/min_requirements.py is available. Older branches don't
-     * have it, so they only get what's hard-coded in the docker files on Linux,
-     * and bare python on other platforms. */
-    public boolean has_min_requirements
-
-    /* Ad hoc overrides for scripts/ci.requirements.txt, used to adjust
-     * requirements on older branches that broke due to updates of the
-     * required packages.
-     * Only used if has_min_requirements is true. */
-    public String python_requirements_override_content
-
-    /* Name of the file containing python_requirements_override_content.
-     * The string is injected into Unix sh and Windows cmd command lines,
-     * so it must not contain any shell escapes or directory separators.
-     * Only used if has_min_requirements is true.
-     * Set to an empty string for convenience if no override is to be
-     * done. */
-    public String python_requirements_override_file
-
-    BranchInfo() {
-        this.all_all_sh_components = [:]
-        this.has_min_requirements = false
-        this.python_requirements_override_content = ''
-        this.python_requirements_override_file = ''
-    }
-}
 
 /* Compute the git object ID of the Dockerfile.
 * Equivalent to the `git hash-object <file>` command. */
